+++ bin/buchta.ts
@@ -3,19 +3,36 @@
 import { readFileSync } from "fs";
 import { readdir } from "fs/promises";
 import { resolve } from "path";
+import { initProject } from "./init";
 
 process.argv = process.argv.slice(2);
+
+export enum BuchtaQuestionType {
+    YES_OR_NO = 16514,
+    TEXT_INPUT,
+    SELECT
+}
+
+export interface BuchtaProjectOption {
+    pretty: string;
+    value?: string;
+    type: BuchtaQuestionType;
+    selectData?: string[];
+}
 
 export class BuchtaCLI {
     cmd: string;
     args: string[];
     projectTemplates = new Map<string, Map<string, string>>();
+    private templateOptions = new Map<string, Map<string, BuchtaProjectOption>>();
+    private templateCallbacks = new Map<string, (opts: Map<string, BuchtaProjectOption>) => void>();
 
     constructor() {
         this.cmd = process.argv[0];
         this.args = process.argv;
+        this.args.shift();
 
-        if (this.args.length == 0 || this.args.includes("-h") || this.args.includes("--help") || /(-h|--h|help)/.exec(this.cmd)) {
+        if (this.args.includes("-h") || this.args.includes("--help") || /(-h|--h|help)/.exec(this.cmd)) {
             this.help();
         }
 
@@ -45,11 +62,15 @@ export class BuchtaCLI {
                     plug.call(this);
                 })
             }
+
+            if (this.cmd == "init") {
+                this.init();
+            }
         })();
     }
 
     init() {
-
+        initProject.call(this);
     }
 
     help() {
@@ -57,7 +78,7 @@ export class BuchtaCLI {
         console.log('Commands:');
         const msgs = [
             "  help\t\tDisplay this message",
-            "  init\t\tCreate new buchta project",
+            "  init\t\t(name)\n\t\tCreate new buchta project",
             "  serve\t\tStart web server",
             "  build\t\tExport your web application",
             "  add\t\t(template|plugin|get|post|put|delete|middleware name)\n\t\tCreate new file from type template",
@@ -74,6 +95,22 @@ export class BuchtaCLI {
             return dirent.isDirectory() ? this.getFiles(res) : res;
         }));
         return Array.prototype.concat(...files);
+    }
+
+    getTemplateOptions(name: string) {
+        return this.templateOptions.get(name) || new Map<string, string>();
+    }
+    
+    setTemplateOptions(name: string, options: Map<string, BuchtaProjectOption>) {
+        this.templateOptions.set(name,  options);
+    }
+
+    getTemplateCallback(name: string) {
+        return this.templateCallbacks.get(name);
+    }
+
+    setTemplateCallback(name: string, callback: (opts: Map<string, BuchtaProjectOption>) => void) {
+        this.templateCallbacks.set(name, callback);
     }
 }