+++ index.d.ts
@@ -2,8 +2,8 @@ import { Buchta, get_version } from "./src/buchta";
 import { BuchtaRequest } from "./src/request";
 import { BuchtaResponse } from "./src/response";
 import { BuchtaSubrouter } from "./src/utils/subrouter";
-import { BuchtaCLI } from "./bin/buchta";
+import { BuchtaCLI, BuchtaQuestionType, BuchtaProjectOption } from "./bin/buchta";
 
 declare module "buchta" {
-    export { Buchta, get_version, BuchtaRequest, BuchtaResponse, BuchtaSubrouter, BuchtaCLI };
+    export { Buchta, get_version, BuchtaRequest, BuchtaResponse, BuchtaSubrouter, BuchtaCLI, BuchtaQuestionType, BuchtaProjectOption };
 }